Consider 64-Bit DataTypes in GetSizeOfAttribute
